refactor(app): move app into its own module
