+++ src/main.rs
@@ -1,6 +1,6 @@
-use color_eyre::eyre::WrapErr;
-use ratatui::crossterm::event::{Event, KeyCode, KeyEvent, KeyEventKind, read};
-use ratatui::{DefaultTerminal, Frame};
+mod app;
+
+use app::App;
 
 fn main() -> color_eyre::Result<()> {
     let _guard = sentry::init((
@@ -18,69 +18,4 @@ fn main() -> color_eyre::Result<()> {
     let result = App::default().run(&mut terminal);
     ratatui::restore();
     result
-}
-
-#[derive(Debug)]
-pub struct App {
-    // TODO: collect text input for search
-    // query: String,
-    running: bool,
-}
-
-impl Default for App {
-    fn default() -> Self {
-        Self {
-            // TODO: collect text input for search
-            // query: String::new(),
-            running: true,
-        }
-    }
-}
-
-impl App {
-    pub fn run(&mut self, terminal: &mut DefaultTerminal) -> color_eyre::Result<()> {
-        while self.running {
-            terminal.draw(|frame| self.draw(frame))?;
-            self.handle_events().wrap_err("handle_events failed")?;
-        }
-        Ok(())
-    }
-
-    fn draw(&self, frame: &mut Frame) {
-        frame.render_widget("hello world", frame.area());
-    }
-
-    fn handle_events(&mut self) -> color_eyre::Result<()> {
-        match read()? {
-            Event::Key(key_ev) if key_ev.kind == KeyEventKind::Press => self
-                .handle_key_event(key_ev)
-                .wrap_err_with(|| format!("handle_key_event failed with:\n{key_ev:#?}")),
-            _ => Ok(()),
-        }
-    }
-
-    fn handle_key_event(&mut self, ev: KeyEvent) -> color_eyre::Result<()> {
-        match ev.code {
-            KeyCode::Char('q') => self.quit(),
-            //TODO: add other keybinds
-            _ => {}
-        }
-        Ok(())
-    }
-
-    fn quit(&mut self) {
-        self.running = false;
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn app_quits_on_key_q() {
-        let mut app = App::default();
-        app.handle_key_event(KeyCode::Char('q').into()).unwrap();
-        assert!(!app.running)
-    }
 }
